\name{Abiotic_data_sample}
\docType{data}
\alias{Abiotic_data_sample}
\title{Dataset with abiotic data and sample names}
\description{Example dataset containing normalized cell numbers and abiotic parameters used for the NMDS plot.}
\usage{data(Abiotic_data_sample)}
\format{Data frame with 10 observations of 44 variables.}
\keyword{datasets}