\name{Cell_number_sample}
\docType{data}
\alias{Cell_number_sample}
\title{Dataset with relative measured cell numbers and sample names}
\description{Example dataset of relative measured cell numbers with sample names.}
\usage{data(Cell_number_sample)}
\format{Data frame with 10 observations of 31 variables.}
\keyword{datasets}