\name{Corr_data_sample}
\docType{data}
\alias{Corr_data_sample}
\title{Dataset with correlation data and sample names}
\description{Example dataset containing relative measured cell numbers and abiotic parameters for correlation analysis.}
\usage{data(Corr_data_sample)}
\format{Data frame with 10 observations of 44 variables.}
\keyword{datasets}
