\name{flowCyBar-package}
\alias{flowCyBar}
\docType{package}
\title{
Analyze flow cytometric data
}
\description{
A package to analyze flow cytometric data
}
\details{
\tabular{ll}{
Package: \tab CyBar\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-14\cr
License: \tab GPL-2\cr
}
Cell_number
Cell_number_nmds
Cell_number_sample
Corr_data
Corr_data_sample
correlation
flowCyBar
cybar_plot
nmds
normalize
Normalized_first
Normalized_mean
Normalized_nmds
}
\author{Joachim Schumann \email{joachim.schumann@ufz.de}, Christin Koch \email{christin.koch@ufz.de}, Susanne G端nther \email{susanne.g端nther@ufz.de}, Ingo Fetzer \email{info.fetzer@stockholmresilience.su.se}, Susann M端ller \email{susann.m端ller@ufz.de}
}
%\references{

%}
%\keyword{ package 

%}
%\seealso{

%}
%\examples{

%}
